+++ cmd/key_list.go
@@ -79,6 +79,8 @@ func BuildCmdListKeys(w io.Writer, handler ListKeysHandler, rf *RootFlags) *cobr
 		"Path to the file containing the wallet's passphrase",
 	)
 
+	autoCompleteWallet(cmd, rf.Home)
+
 	return cmd
 }